# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Система управления задачами - API
  description: |
    Микросервисная архитектура для управления пользователями и заказами
    в строительных проектах. ООО "СистемаКонтроля".
  version: 1.0.0
  contact:
    name: ООО СистемаКонтроля
    email: support@sistemakontrolya.ru
servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: http://test-server:8000/api/v1
    description: Test server
  - url: https://api.production.com/api/v1
    description: Production server
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT токен, полученный при входе
  schemas:
    Error:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Некорректные данные"
...
    OrderItem:
      type: object
      required:
        - product
        - quantity
        - price
      properties:
        product:
          type: string
          example: "Кирпич красный"
        quantity:
          type: integer
          minimum: 1
          example: 1000
        price:
          type: number
          minimum: 0
          example: 15.50
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        items:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
        status:
          type: string
          enum: [created, in_progress, completed, cancelled]
          example: "created"
        totalAmount:
          type: number
          example: 15500.00
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    Pagination:
      type: object
      properties:
        page:
          type: integer
          example: 1
        limit:
          type: integer
          example: 10
        total:
          type: integer
          example: 100
        totalPages:
          type: integer
          example: 10
...
paths:
...
  /orders:
    post:
      summary: Создание нового заказа
      description: Создает новый заказ для текущего пользователя
      tags: [Orders]
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - items
              properties:
                items:
                  type: array
                  minItems: 1
                  items:
                    $ref: '#/components/schemas/OrderItem'
            example:
              items:
                - product: "Цемент М500"
                  quantity: 50
                  price: 350.00
                - product: "Песок речной"
                  quantity: 100
                  price: 80.00
      responses:
        '201':
          description: Заказ создан
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Не авторизован
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    get:
      summary: Список заказов текущего пользователя
      description: Возвращает заказы текущего пользователя с пагинацией и сортировкой
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 10
        - name: sortBy
          in: query
          description: Поле для сортировки
          schema:
            type: string
            enum: [createdAt, updatedAt, totalAmount, status]
            default: createdAt
        - name: sortOrder
          in: query
          description: Порядок сортировки
          schema:
            type: string
            enum: [asc, desc]
            default: desc
      responses:
        '200':
          description: Список заказов
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      orders:
                        type: array
                        items:
                          $ref: '#/components/schemas/Order'
                      pagination:
                        $ref: '#/components/schemas/Pagination'
  /orders/{id}:
    get:
      summary: Получение заказа по ID
      description: Возвращает детали заказа. Пользователи видят только свои заказы, админы - все.
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: UUID заказа
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Информация о заказе
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '403':
          description: Недостаточно прав (попытка доступа к чужому заказу)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Отмена заказа
      description: |
        Отменяет заказ (устанавливает статус cancelled). 
        Нельзя отменить завершенный заказ.
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Заказ отменен
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Невозможно отменить заказ (например, уже завершен)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{id}/status:
    patch:
      summary: Обновление статуса заказа
      description: Изменяет статус заказа. Доступно владельцу заказа и администраторам.
      tags: [Orders]
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - status
              properties:
                status:
                  type: string
                  enum: [created, in_progress, completed, cancelled]
            example:
              status: "in_progress"
      responses:
        '200':
          description: Статус обновлен
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    $ref: '#/components/schemas/Order'
        '400':
          description: Недопустимый статус
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Недостаточно прав
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Заказ не найден
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...